move page contents
