+++ src/app/about/page.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { AboutPageContents } from "../components/AboutPageContents"
+import { AboutPageContents } from "./components/AboutPageContents"
 
 export default function AboutPage() {
   return <AboutPageContents />
