+++ src/app/contact/page.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { ContactPageContents } from "../components/ContactPageContents"
+import { ContactPageContents } from "./components/ContactPageContents"
 
 export default function ContactPage() {
   return <ContactPageContents />
